feat(evidence): split writer.rs into manifest, limits, verify, write (E2)
- Split writer.rs into manifest.rs, limits.rs, verify.rs, write.rs
- Façade re-exports: XAssayExtension, BundleProvenance for one-stop API
- Add public_api_smoke test for bundle façade compile-safety
- Evidence export: populate bundle_provenance via ProvenanceInput (Phase 2)
- ADR-025-E2: document bundle_digest as logical digest (run_root, algorithms, files)
- Update imports across lint, diff, assay-sim, tests, benches
- TODOS.md: record writer-split as done
